updated html_parser for benchmarking and re-orged directories
