+++ Gemfile
@@ -1,3 +1,4 @@
 source "https://rubygems.org"
 
 gem 'nokogiri'
+gem 'benchmark'
